+++ Download/DownloadTask.swift
@@ -59,68 +59,34 @@ public final class DownloadTask {
         self.fairplayRequester = fairplayRequester
     }
     
-    // Controls
+    // MARK: Controls
     public func resume() {
         // AVAssetDownloadTask provides the ability to resume previously stopped downloads under certain circumstances. To do so, simply instantiate a new AVAssetDownloadTask with an AVURLAsset instantiated with a file NSURL pointing to the partially downloaded bundle with the desired download options, and the download will continue restoring any previously downloaded data. FPS keys remain encrypted in persisted form during this process.
         guard let task = task else {
-            print("⚠️ Resume called on DownloadTask in preparation")
+            createNewTask(with: configuration)
             return
         }
         task.resume()
         onResumed(self)
     }
     
-    public func prepare() { // TODO: Possibly remove this
-        if task != nil {
-            print("⚠️ Prepare called on DownloadTask with an attached AVAssetDownloadTask")
-            print("👍 DownloadTask prepared")
-            onPrepared(self)
-            
-            if autoStart {
-                resume()
-            }
-            return
-        }
-//        restoreSuspendedTask(for: configuration) { [weak self] success in
-//            guard let weakSelf = self else { return }
-//            if !success {
-//                weakSelf.createNewTask(with: weakSelf.configuration)
-//            }
-//        }
-        createNewTask(with: configuration)
-    }
-    
-//    private func restoreSuspendedTask(for configuration: Configuration, callback: @escaping (Bool) -> Void) {
-//        print("♻️ Attempting to restore previous DownloadTask for: \(configuration.assetId)")
-//        sessionManager.task(assetId: configuration.assetId) { [weak self] task in
-//            guard let weakSelf = self else {
-//                return
-//            }
-//            guard let task = task else {
-//                print("⚠️ No DownloadTask to restore for: \(weakSelf.configuration.assetId)")
-//                callback(false)
-//                return
-//            }
-//            print("♻️ Restored DownloadTask associated with request for: \(weakSelf.configuration.assetId)")
-//            
-//            weakSelf.finalizePreparation(of: task)
-//        }
-//    }
-    
     private func createNewTask(with configuration: Configuration) {
         print("✅ Creating new DownloadTask for: \(configuration.assetId)")
         // Create a fresh task
         let options = requiredBitrate != nil ? [AVAssetDownloadTaskMinimumRequiredMediaBitrateKey: requiredBitrate!] : nil
         
-        startTask(with: options) { error in
+        startTask(with: options) { task, error in
             if let error = error {
                 onError(self, configuration.destination, error)
                 return
             }
+            task!.resume()
+            onStarted(self)
+            onResumed(self)
         }
     }
     
-    internal func startTask(with options: [String: Any]?, callback: (DownloadError?) -> Void) {
+    internal func startTask(with options: [String: Any]?, callback: (AVAssetDownloadTask? ,DownloadError?) -> Void) {
             if #available(iOS 10.0, *) {
                 guard let task = sessionManager
                     .session
@@ -129,12 +95,12 @@ public final class DownloadTask {
                                            assetArtworkData: configuration.artwork,
                                            options: options) else {
                                             // This method may return nil if the AVAssetDownloadURLSession has been invalidated.
-                                            callback(.downloadSessionInvalidated)
+                                            callback(nil,.downloadSessionInvalidated)
                                             return
                 }
                 task.taskDescription = configuration.assetId
                 finalizePreparation(of: task)
-                callback(nil)
+                callback(task,nil)
             }
             else {
                 guard let destination = configuration.destination else {
@@ -147,12 +113,12 @@ public final class DownloadTask {
                                            destinationURL: destination,
                                            options: options) else {
                                             // This method may return nil if the URLSession has been invalidated
-                                            callback(.downloadSessionInvalidated)
+                                            callback(nil,.downloadSessionInvalidated)
                                             return
                 }
                 task.taskDescription = configuration.assetId
                 finalizePreparation(of: task)
-                callback(nil)
+                callback(task,nil)
             }
     }
     
@@ -173,11 +139,6 @@ public final class DownloadTask {
         sessionManager.delegate[task] = self
         
         print("👍 DownloadTask prepared")
-        onPrepared(self)
-        
-        if autoStart {
-            resume()
-        }
     }
     
     public func suspend() {
@@ -194,13 +155,6 @@ public final class DownloadTask {
         
         guard let task = self.task else { return }
         task.cancel()
-//        switch task.state {
-//        case .suspended:
-//            task.resume()
-//            task.cancel()
-//        default:
-//            task.cancel()
-//        }
         
         // NOTE: `onCanceled` called once `didCompleteWithError` delegate methods is triggered
     }
@@ -214,13 +168,6 @@ public final class DownloadTask {
     @discardableResult
     public func use(bitrate: Int64?) -> DownloadTask {
         requiredBitrate = bitrate
-        return self
-    }
-    
-    fileprivate var autoStart: Bool = true
-    @discardableResult
-    public func should(autoStart: Bool = true) -> DownloadTask {
-        self.autoStart = autoStart
         return self
     }
     
@@ -248,7 +195,6 @@ public final class DownloadTask {
     //    }
     
     // MARK: DownloadEventPublisher
-    internal var onPrepared: (DownloadTask) -> Void = { _ in }
     internal var onStarted: (DownloadTask) -> Void = { _ in }
     internal var onSuspended: (DownloadTask) -> Void = { _ in }
     internal var onResumed: (DownloadTask) -> Void = { _ in }
@@ -284,12 +230,6 @@ extension DownloadTask {
 extension DownloadTask: DownloadEventPublisher {
     public typealias DownloadEventProgress = Progress
     public typealias DownloadEventError = DownloadError
-    
-    @discardableResult
-    public func onPrepared(callback: @escaping (DownloadTask) -> Void) -> DownloadTask {
-        onPrepared = callback
-        return self
-    }
     
     @discardableResult
     public func onStarted(callback: @escaping (DownloadTask) -> Void) -> DownloadTask {