+++ CHANGELOG.md
@@ -10,6 +10,8 @@ NO RELEASE YET
 * `EMP-10327` Initial download functionality added.
 * `EMP-10445` Download functionality with Session management.
 * `EMP-10474` Persist and retrieve `FairPlay` content keys in download and offline scenarios.
+* `EMP-10478` Preparation of `DownloadTask`s now occur once `resume()` is called.
+    
 
 #### Known Limitations
 * Apple confirmed `iOS` `10.3 beta 3` fixes a bug (possibly introduced in `iOS 10.2.1`) where *suspending/resuming* a download multiple times (*3*) causes the `AVAssetDownloadTask` to enter a freezed state. More importantly, once this occurs it is impossible to start new downloading tasks unless a the affected device is *restarted*.